+++ src/services/SendoutsService.php
@@ -154,14 +154,13 @@ class SendoutsService extends Component
             /** @var SendoutElement $sendout */
             foreach ($sendouts as $sendout) {
                 if ($sendout->getCanSendNow()) {
-                    $job = new SendoutJob([
-                        'sendoutId' => $sendout->id,
-                        'title' => SendoutHelper::encodeEmojis($sendout->title),
-                    ]);
-
                     Queue::push(
-                        job: $job,
+                        job: new SendoutJob([
+                            'sendoutId' => $sendout->id,
+                            'title' => SendoutHelper::encodeEmojis($sendout->title),
+                        ]),
                         priority: Campaign::$plugin->settings->sendoutJobPriority,
+                        ttr: Campaign::$plugin->settings->sendoutJobTtr,
                         queue: Campaign::$plugin->queue,
                     );